+++ watch.go
@@ -9,11 +9,8 @@ import (
 	"fmt"
 	"log/slog"
 	"reflect"
-	"slices"
 	"sync"
 	"time"
-
-	"github.com/nil-go/konf/internal/maps"
 )
 
 // Watch watches and updates configuration when it changes.
@@ -24,39 +21,31 @@ import (
 func (c *Config) Watch(ctx context.Context) error { //nolint:cyclop,funlen,gocognit
 	c.nocopy.Check()
 
-	if hasWatcher := slices.ContainsFunc(c.providers, func(provider *provider) bool {
-		_, ok := provider.loader.(Watcher)
-
-		return ok
-	}); !hasWatcher {
-		return nil
-	}
-
 	if watched := c.watched.Swap(true); watched {
-		c.log(ctx, slog.LevelWarn, "Config has been watched, call Watch again has no effects.")
+		c.log(ctx, slog.LevelWarn, "Config has been watched, call Watch more than once has no effects.")
 
 		return nil
 	}
 
 	ctx, cancel := context.WithCancelCause(ctx)
 	defer cancel(nil)
-	var waitGroup sync.WaitGroup
 
 	// Start a goroutine to update the configuration while it has changes from watchers.
 	onChangesChannel := make(chan []func(*Config), 1)
 	defer close(onChangesChannel)
+
+	var waitGroup sync.WaitGroup
 	waitGroup.Add(1)
 	go func() {
 		defer waitGroup.Done()
 
 		for {
 			select {
+			case <-ctx.Done():
+				return
+
 			case onChanges := <-onChangesChannel:
-				values := make(map[string]any)
-				for _, w := range c.providers {
-					maps.Merge(values, *w.values.Load())
-				}
-				c.values.Store(&values)
+				c.providers.sync()
 				c.log(ctx, slog.LevelDebug, "Configuration has been updated with change.")
 
 				if len(onChanges) > 0 {
@@ -79,25 +68,19 @@ func (c *Config) Watch(ctx context.Context) error { //nolint:cyclop,funlen,gocog
 							if errors.Is(tctx.Err(), context.DeadlineExceeded) {
 								c.log(
 									ctx, slog.LevelWarn,
-									"Configuration has not been fully applied to onChanges due to timeout."+
+									"Configuration has not been fully applied to onChanges in one minute."+
 										" Please check if the onChanges is blocking or takes too long to complete.",
 								)
 							}
 						}
 					}()
 				}
-
-			case <-ctx.Done():
-				return
 			}
 		}
 	}()
 
 	// Start a watching goroutine for each watcher registered.
-	c.providersMutex.RLock()
-	for i := range c.providers {
-		provider := c.providers[i] // Use pointer for later modification.
-
+	c.providers.traverse(func(provider *provider) {
 		if watcher, ok := provider.loader.(Watcher); ok {
 			waitGroup.Add(1)
 			go func(ctx context.Context) {
@@ -106,24 +89,11 @@ func (c *Config) Watch(ctx context.Context) error { //nolint:cyclop,funlen,gocog
 				onChange := func(values map[string]any) {
 					c.transformKeys(values)
 					oldValues := *provider.values.Swap(&values)
-
-					// Find the onChanges should be triggered.
-					onChanges := func() []func(*Config) {
-						c.onChangesMutex.RLock()
-						defer c.onChangesMutex.RUnlock()
-
-						var callbacks []func(*Config)
-						for path, onChanges := range c.onChanges {
-							oldVal := c.sub(oldValues, path)
-							newVal := c.sub(values, path)
-							if !reflect.DeepEqual(oldVal, newVal) {
-								callbacks = append(callbacks, onChanges...)
-							}
-						}
-
-						return callbacks
-					}
-					onChangesChannel <- onChanges()
+					onChangesChannel <- c.onChanges.get(
+						func(path string) bool {
+							return !reflect.DeepEqual(c.sub(oldValues, path), c.sub(values, path))
+						},
+					)
 
 					c.log(ctx, slog.LevelInfo,
 						"Configuration has been changed.",
@@ -137,8 +107,7 @@ func (c *Config) Watch(ctx context.Context) error { //nolint:cyclop,funlen,gocog
 				}
 			}(ctx)
 		}
-	}
-	c.providersMutex.RUnlock()
+	})
 	waitGroup.Wait()
 
 	if err := context.Cause(ctx); err != nil && !errors.Is(err, ctx.Err()) {
@@ -156,27 +125,52 @@ func (c *Config) Watch(ctx context.Context) error { //nolint:cyclop,funlen,gocog
 // The register function must be non-blocking and usually completes instantly.
 // If it requires a long time to complete, it should be executed in a separate goroutine.
 //
-// This method is concurrency-safe.
+// This method is concurrent-safe.
 func (c *Config) OnChange(onChange func(*Config), paths ...string) {
 	if onChange == nil {
 		return // Do nothing is onchange is nil.
 	}
 	c.nocopy.Check()
 
+	if !c.caseSensitive {
+		for i := range paths {
+			paths[i] = defaultKeyMap(paths[i])
+		}
+	}
+	c.onChanges.register(onChange, paths)
+}
+
+type onChanges struct {
+	subscribers map[string][]func(*Config)
+	mutex       sync.RWMutex
+}
+
+func (o *onChanges) register(onChange func(*Config), paths []string) {
+	o.mutex.Lock()
+	defer o.mutex.Unlock()
+
 	if len(paths) == 0 {
 		paths = []string{""}
 	}
 
-	c.onChangesMutex.Lock()
-	defer c.onChangesMutex.Unlock()
-
-	if c.onChanges == nil { // To support zero Config
-		c.onChanges = make(map[string][]func(*Config))
+	if o.subscribers == nil {
+		o.subscribers = make(map[string][]func(*Config))
 	}
 	for _, path := range paths {
-		if !c.caseSensitive {
-			path = defaultKeyMap(path)
-		}
-		c.onChanges[path] = append(c.onChanges[path], onChange)
+		o.subscribers[path] = append(o.subscribers[path], onChange)
 	}
+}
+
+func (o *onChanges) get(filter func(string) bool) []func(*Config) {
+	o.mutex.RLock()
+	defer o.mutex.RUnlock()
+
+	var callbacks []func(*Config)
+	for path, subscriber := range o.subscribers {
+		if filter(path) {
+			callbacks = append(callbacks, subscriber...)
+		}
+	}
+
+	return callbacks
 }